(PC-28481)[PRO] refactor: move provider files to VenueSettings folder
