--- pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/AddVenueProviderButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AddVenueProviderButton'
