--- pro/src/pages/VenueCreation/OffersSynchronization/index.ts
@@ -1,1 +1,0 @@
-export { default as OffersSynchronization } from './OffersSynchronization'
