+++ pro/src/pages/VenueSettings/VenueProvidersManager/AllocineProviderForm/__specs__/AllocineProviderForm.spec.tsx
@@ -7,7 +7,8 @@ import { SynchronizationEvents } from 'core/FirebaseEvents/constants'
 import * as useAnalytics from 'hooks/useAnalytics'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
-import AllocineProviderForm, {
+import {
+  AllocineProviderForm,
   AllocineProviderFormProps,
 } from '../AllocineProviderForm'
