+++ pro/src/pages/VenueSettings/VenueProvidersManager/CinemaProviderForm/__specs__/CinemaProviderForm.spec.tsx
