+++ pro/src/pages/VenueSettings/VenueProvidersManager/StocksProviderForm/StocksProviderForm.module.scss
