+++ pro/src/pages/VenueSettings/VenueProvidersManager/VenueProviderList/CinemaProviderFormDialog.scss
