+++ pro/src/pages/VenueSettings/VenueProvidersManager/VenueProviderList/DeleteVenueProviderDialog.module.scss
